Making SEB-configuration more granularly
There is 2 new configuration options:
SEB_WHITELIST_PATHS = []
SEB_BLACKLIST_CHAPTERS = []

SEB_WHITELIST_PATHS is a list of paths that can be accessed, it may contain one of the following values: 'wiki', 'course-outline', 'courseware', 'progress', 'discussion'; plus any other url name (e.g. useful for custom tabs)

SEB_BLACKLIST_CHAPTERS is a list of chapters that are blocked even when 'courseware' has been white-listed; e.g. useful to block the "exam" chapter but not the rest.

Right now those 2 configurations are global and not in a course-basis; may be better to agree on the current state before making it so.